+++ genlib/osciln.c
@@ -1,4 +1,4 @@
-float osciln(float amp, float si, float *farray, int len, float *phs)
+float osciln(float amp, float si, double *farray, int len, float *phs)
 {
 	register int i =  *phs;   
 	*phs += si;
